separate region-list
