+++ catalystwan/endpoints/configuration/policy/list/region.py
@@ -4,17 +4,9 @@
 from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
-from catalystwan.models.policy.lists import RegionList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.region import RegionList, RegionListEditPayload, RegionListInfo
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class RegionListEditPayload(RegionList, PolicyListId):
-    pass
-
-
-class RegionListInfo(RegionList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicyRegionList(APIEndpoints):
